requirements.txt
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.1.0
certifi==2024.12.14
frozenlist==1.5.0
greenlet==3.1.1
hiredis==3.1.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
psycopg2-binary==2.9.10
pyaes==1.6.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
redis==5.0.8
SQLAlchemy==2.0.37
TgCrypto==1.2.5
typing_extensions==4.12.2
yarl==1.18.3
